+++ nata-monkey/src/main/java/com/nata/action/StartAppAction.java
@@ -1,6 +1,6 @@
 package com.nata.action;
 
-import com.nata.cmd.AdbDevice;
+import com.nata.AdbDevice;
 
 /**
  * Author: Calvin Meng