+++ nata-monkey/src/main/java/com/nata/runners/DFSRunner.java
@@ -1,9 +1,8 @@
 package com.nata.runners;
 
-import com.nata.cmd.AdbDevice;
+import com.nata.AdbDevice;
 import com.nata.monkeys.AbstractMonkey;
 import com.nata.monkeys.DfsMonkey;
-import com.nata.monkeys.PureRandomMonkey;
 
 /**
  * Author: Calvin Meng