+++ nata-monkey/src/main/java/com/nata/runners/RMRunner.java
@@ -1,6 +1,6 @@
 package com.nata.runners;
 
-import com.nata.cmd.AdbDevice;
+import com.nata.AdbDevice;
 import com.nata.monkeys.RandomMonkey;
 
 /**
@@ -11,7 +11,7 @@ import com.nata.monkeys.RandomMonkey;
 public class RMRunner {
     public static void main(String[] args) {
 //      String pkg = "com.zhihu.android";
-//      String act = ".app.ui.activity.MainActivity";
+//      String act = ".App.ui.activity.MainActivity";
 
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";